+++ main/src/main/java/net/oneandone/stool/cli/Start.java
@@ -16,7 +16,7 @@
 package net.oneandone.stool.cli;
 
 import net.oneandone.inline.ArgumentException;
-import net.oneandone.stool.configuration.Until;
+import net.oneandone.stool.configuration.Expire;
 import net.oneandone.stool.locking.Mode;
 import net.oneandone.stool.stage.Stage;
 import net.oneandone.stool.util.Files;
@@ -74,7 +74,7 @@ public class Start extends StageCommand {
 
         serviceWrapperOpt(stage);
         download = tomcatOpt(stage.config().tomcatVersion);
-        checkUntil(stage.config().until);
+        checkExpired(stage.config().expire);
         if (session.configuration.committed) {
             if (!stage.isCommitted()) {
                 throw new IOException("It's not allowed to start stages with local modifications.\n"
@@ -411,9 +411,9 @@ public class Start extends StageCommand {
         return result.toString();
     }
 
-    private void checkUntil(Until until) {
-        if (until.isExpired()) {
-            throw new ArgumentException("Stage expired " + until + ". To start it, you have to adjust the 'until' date.");
+    private void checkExpired(Expire expire) {
+        if (expire.isExpired()) {
+            throw new ArgumentException("Stage expired " + expire + ". To start it, you have to adjust the 'expire' date.");
         }
     }
 }