+++ main/src/main/java/net/oneandone/stool/setup/Lib.java
@@ -230,6 +230,9 @@ public class Lib {
             String suffixesRename() {
                 return "urls";
             }
+            String untilRename() {
+                return "expire";
+            }
         };
     }
 }